btrfs-progs: Force 'btrfs dev scan' not using old libblkid cache
'btrfs scan' uses libblkid to scan devices by default, and libblkid uses
cache to reduce the probe.

But if operations below is done in less than 2 seconds(BLKID_PROBE_MIN),
'btrfs scan' will still use the uncorrect cache and scan on the deleted device.
0. /dev/sda[1-4] mounted on /mnt using single data/metadata
1. btrfs dev scan
2. btrfs dev del /dev/sda3 /mnt
3. btrfs dev scan

Since the cache made by step 1 is still validated, step 3 will use the
cache and consider /dev/sda3 as a btrfs filesystem and try to scan it.
But the superblock(at least the first one) is wiped and failed to scan,
a error message, which can be avoided and is unneeded, is output.

This patch will force scan_for_btrfs() not to use cache to avoid the
problem.

Reported-by: Tsutomu Itoh <t-itoh@jp.fujitsu.com>
Singed-off-by: Qu Wenruo <quwenruo@cn.fujitsu.com>

Signed-off-by: David Sterba <dsterba@suse.cz>